refactor: move tag formatting logic to config package
