+++ internal/librarian/commit_version_analyzer_test.go
@@ -75,49 +75,6 @@ func TestShouldExclude(t *testing.T) {
 	}
 }
 
-func TestFormatTag(t *testing.T) {
-	t.Parallel()
-	for _, test := range []struct {
-		name    string
-		library *config.LibraryState
-		want    string
-	}{
-		{
-			name: "default format",
-			library: &config.LibraryState{
-				ID:      "google.cloud.foo.v1",
-				Version: "1.2.3",
-			},
-			want: "google.cloud.foo.v1-1.2.3",
-		},
-		{
-			name: "custom format",
-			library: &config.LibraryState{
-				ID:        "google.cloud.foo.v1",
-				Version:   "1.2.3",
-				TagFormat: "v{version}-{id}",
-			},
-			want: "v1.2.3-google.cloud.foo.v1",
-		},
-		{
-			name: "custom format -- version only",
-			library: &config.LibraryState{
-				ID:        "google.cloud.foo.v1",
-				Version:   "1.2.3",
-				TagFormat: "v{version}",
-			},
-			want: "v1.2.3",
-		},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			got := formatTag(test.library.TagFormat, test.library.ID, test.library.Version)
-			if got != test.want {
-				t.Errorf("formatTag() = %q, want %q", got, test.want)
-			}
-		})
-	}
-}
-
 func TestGetConventionalCommitsSinceLastRelease(t *testing.T) {
 	t.Parallel()
 	pathAndMessages := []pathAndMessage{
