+++ internal/librarian/release_notes.go
@@ -269,8 +269,8 @@ func formatReleaseNotes(state *config.LibrarianState, ghRepo *github.Repository)
 func formatLibraryReleaseNotes(library *config.LibraryState, ghRepo *github.Repository) *releaseNoteSection {
 	// The version should already be updated to the next version.
 	newVersion := library.Version
-	newTag := formatTag(library.TagFormat, library.ID, newVersion)
-	previousTag := formatTag(library.TagFormat, library.ID, library.PreviousVersion)
+	newTag := config.FormatTag(library.TagFormat, library.ID, newVersion)
+	previousTag := config.FormatTag(library.TagFormat, library.ID, library.PreviousVersion)
 
 	commitsByType := make(map[string][]*conventionalcommits.ConventionalCommit)
 	for _, commit := range library.Changes {
